+++ config/default.js
@@ -1090,7 +1090,8 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 	 */
 	typograf() {
 		return {
-			locale: this.locale
+			// Typograf dont support en locale
+			locale: this.locale === 'en' ? 'en-US' : this.locale
 		};
 	},