+++ src/form/b-select-date/b-select-date.i18n/en.js
@@ -1,0 +1,24 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+module.exports = {
+	'b-select-date': {
+		January: 'January',
+		February: 'February',
+		March: 'March',
+		April: 'April',
+		May: 'May',
+		June: 'June',
+		July: 'July',
+		August: 'August',
+		September: 'September',
+		October: 'October',
+		November: 'November',
+		December: 'December'
+	}
+};